Changed architecture to pass data from php to java
The original design was to use a named pipe in linux to pass
data from the php post to the queue, however java blocks the thread
while it tries to read from the named pipe. There really is no
good way around this.

Thus, I am making a file that all the commands get wrtten to and read from.
The queue will read each line, add it to the queue, then delete the file.

There is also a file.lock made that acts as a semaphore between the two
applications. I did this because I'm not 100% sure how file locks will behave
in both php and java. This extra file gives me control over access to the
pipe of information. When an application is using it it will make the file.lock
and when it's done it removes it.